requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
aiosqlite==0.20.0
cryptography==43.0.0
apscheduler==3.10.4
